# server-skeleton/docs/swagger.yaml
definitions:
  page.PageItemResultDto:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  page.RequestPageDTO:
    properties:
      created_at:
        example: "2022-01-01T00:00:00Z"
        type: string
      deleted_at:
        example: "2022-01-01T00:00:00Z"
        type: string
      name:
        example: Some Page
        type: string
      updated_at:
        example: "2022-01-01T00:00:00Z"
        type: string
    type: object
  page.SuccessResponseDto:
    properties:
      message:
        type: string
    type: object
info:
  contact: {}
paths:
  /page:
    get:
      consumes:
      - application/json
      description: Getting pages
      parameters:
      - collectionFormat: csv
        description: Name
        in: query
        items:
          type: string
        name: names
        type: array
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: cursor (las id uuid)
        in: query
        name: cursor
        type: string
      - description: lastTimestamp
        in: query
        name: lastTimestamp
        type: string
      - collectionFormat: csv
        description: 'Filter created_at Like min-max (пример: 2024-01-01_2024-12-31)'
        in: query
        items:
          type: string
        name: orders[created_at]
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/page.RequestPageDTO'
            type: array
      tags:
      - pages
    post:
      consumes:
      - application/json
      description: Create page
      parameters:
      - description: Sent date
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/page.RequestPageDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/page.SuccessResponseDto'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create page
      tags:
      - pages
  /page/{id}:
    delete:
      consumes:
      - application/json
      description: Deleting page by id
      parameters:
      - description: Page id (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/page.RequestPageDTO'
            type: array
      tags:
      - page
    get:
      consumes:
      - application/json
      description: Getting page by id
      parameters:
      - description: Page id (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/page.RequestPageDTO'
            type: array
      tags:
      - page
    patch:
      consumes:
      - application/json
      description: Update only sent fields
      parameters:
      - description: Page id (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Updated data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/page.RequestPageDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/page.PageItemResultDto'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Patch page
      tags:
      - page
    put:
      consumes:
      - application/json
      description: Update all sent fields
      parameters:
      - description: Page id (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Updated data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/page.RequestPageDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/page.SuccessResponseDto'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Put page
      tags:
      - page
swagger: "2.0"